requirements.txt
streamlit
langchain
langchain-anthropic
langchain-community
langchain-core
anthropic
tavily-python
python-dotenv
PyPDF2
python-docx
SpeechRecognition
